+++ apps/client/src/app/modules/inventory/inventory-position/inventory-position.component.ts
@@ -1,9 +1,9 @@
 import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
 import { ItemSearchResult } from '../../../model/user/inventory/item-search-result';
-import { InventoryFacade } from '../+state/inventory.facade';
 import { ContainerType } from '../../../model/user/inventory/container-type';
 import { ReplaySubject } from 'rxjs';
-import { map, switchMap } from 'rxjs/operators';
+import { filter, map, switchMap } from 'rxjs/operators';
+import { InventoryService } from '../inventory.service';
 
 @Component({
   selector: 'app-inventory-position',
@@ -21,6 +21,7 @@ export class InventoryPositionComponent {
   }
 
   public display$ = this.item$.pipe(
+    filter(item => !!item),
     switchMap(item => {
       return this.inventoryFacade.getPosition(item).pipe(
         map(position => {
@@ -43,7 +44,7 @@ export class InventoryPositionComponent {
     })
   );
 
-  constructor(private inventoryFacade: InventoryFacade) {
+  constructor(private inventoryFacade: InventoryService) {
   }
 
   private getEmptyInventory(item: ItemSearchResult): { size: [number, number, number], array: boolean[][][] } {