+++ apps/client/src/environments/patch-notes.ts
@@ -1,13 +1,16 @@
 export const patchNotes = `### Bug Fixes
 
-* **desktop:** fixed random inventory resets for some items.
-* **layout:** lowered npc breakdown priority for beast tribe npcs.
-* **leveling-equipment:** fixed multiple issues with gear search system.
-* **list:** crystals cannot be suggested as HQ anymore.
-* **simulator:** fixed Observe combo being broken by skipped actions.
-* **treasure-finder:** fixed wrong mini checkmark position.
+* **desktop:** fixed search sharing button not giving the right url.
+* **leveling-equipment:** masterbook items are no longer being considered.
+* **recipe-finder:** fixed an issue with some recipes not being found properly.
+* **requirements:** fixed inventory position tag not showing properly.
+* **simulator:** fixed observe combo skipping with standard touch.
+* **voyage-tracker:** fixed a issue with voyage tracker not loading in some cases.
 
 
-### Performance Improvements
+### Features
 
-* **collectables:** tabs are now lazy loaded for better page loading perfs.`;
+* **data:** update for cn patch-5.4.
+* **desktop:** complete refactor of inventory tracking for better reliability.
+* **rotation:** added a button to generate macro in the rotation panel.
+* **simulator:** support for CN 5.4.`;